feat:完成docker配置並部署

+++ order-service/src/main/java/com/rafa/order_service/controller/OrderController.java
@@ -6,7 +6,10 @@ import com.rafa.order_service.feign.RoomInterface;
 import com.rafa.order_service.feign.UserInterface;
 import com.rafa.order_service.model.Orders;
 import com.rafa.order_service.model.dto.*;
+import com.rafa.order_service.response.ApiResponse;
 import com.rafa.order_service.service.OrderService;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.amqp.rabbit.annotation.RabbitListener;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -19,6 +22,7 @@ import java.util.List;
 
 @RestController
 @RequestMapping("/member/order")
+@Slf4j
 public class OrderController {
 
     @Autowired
@@ -36,14 +40,15 @@ public class OrderController {
     ObjectMapper objectMapper;
 
     @PostMapping
-    public ResponseEntity<Orders> createOrder(@RequestHeader("Authorization") String jwt, @RequestBody CreateOrderRequest request) {
+    public ResponseEntity<ApiResponse<Orders>> createOrder(@RequestHeader("Authorization") String jwt, @RequestBody CreateOrderRequest request) {
         try {
             UserDto user = objectMapper.convertValue(userInterface.getUserProfile(jwt).getBody().getData(), UserDto.class);
             if (
                     !request.getCheckInDate().isBefore(LocalDate.now()) &&//不可預訂過去日期
                             request.getCheckInDate().isBefore(request.getCheckOutDate()) &&//審查日期合理性
                             orderService.isRoomAvailable(request.getRoomId(), request.getCheckInDate(), request.getCheckOutDate())) { //不可重複預定
-                List<RoomDto> onlyRoom = roomInterface.getRoomCardsByTimeFromRoomIds(List.of(request.getRoomId()), null, null).getBody();
+
+                List<RoomDto> onlyRoom = roomInterface.getRoomCardsByTimeFromRoomIds(List.of(request.getRoomId()), null, null).getBody().getData();
                 RoomDto room = new RoomDto();
                 if (onlyRoom != null)
                     room = onlyRoom.get(0);
@@ -51,45 +56,46 @@ public class OrderController {
                 Orders newOrder = new Orders();
                 if (user != null)
                     newOrder = orderService.createOrder(user, room.getRoomId(), request.getTotalPrice(), room.getRoomName(), !room.getRoomPic().isEmpty() ? room.getRoomPic().get(0) : "", request.getCheckInDate(), request.getCheckOutDate());
-                return new ResponseEntity<>(newOrder, HttpStatus.CREATED);
-            } else throw new RuntimeException("訂房失敗，請重新操作");
+                return ResponseEntity.ok(ApiResponse.success("預訂成功", newOrder));
+            }
+            return ResponseEntity.badRequest().body(ApiResponse.error(400, "訂房資料錯誤，請重新確認"));
         } catch (RuntimeException e) {
             e.printStackTrace();
         }
-        return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
+        return ResponseEntity.badRequest().body(ApiResponse.error(400, "預訂失敗，請重新操作"));
     }
 
     @DeleteMapping("/{orderId}")
-    public ResponseEntity<String> deleteInvalidOrderFromUser(@RequestHeader("Authorization") String jwt, @PathVariable long orderId) {
+    public ResponseEntity<ApiResponse<String>> deleteInvalidOrderFromUser(@RequestHeader("Authorization") String jwt, @PathVariable long orderId) {
         try {
-            Long userId = objectMapper.convertValue(authInterface.findUserIdByJwt(jwt).getBody().getData(), Long.class);
+            Long userId = authInterface.findUserIdByJwt(jwt).getBody().getData();
+            Orders order = orderService.findOrderByOrderId(orderId);
+            if (userId != order.getUserId())
+                return ResponseEntity.badRequest().body(ApiResponse.error(400, "無權限執行此操作"));
             System.out.println("我要刪除order" + orderId);
             if (orderService.deleteInvalidOrderFromUser(userId, orderId))
-
-                return new ResponseEntity<>("成功刪除", HttpStatus.OK);
+                return ResponseEntity.ok(ApiResponse.success("成功刪除", null));
         } catch (Exception e) {
             e.printStackTrace();
         }
-        return new ResponseEntity<>("錯誤", HttpStatus.BAD_REQUEST);
+        return ResponseEntity.badRequest().body(ApiResponse.error(400, "訂單刪除失敗，請重新嘗試"));
     }
 
     @GetMapping("/availability/{roomId}")
-    public ResponseEntity<Boolean> checkRoomAvailable(@PathVariable Long roomId, @RequestParam LocalDate start, @RequestParam LocalDate end) {
+    public ResponseEntity<ApiResponse<Boolean>> checkRoomAvailable(@PathVariable Long roomId, @RequestParam LocalDate start, @RequestParam LocalDate end) {
         try {
-            if (orderService.isRoomAvailable(roomId, start, end))
-                return new ResponseEntity<>(true, HttpStatus.OK);
-            else
-                return new ResponseEntity<>(false, HttpStatus.OK);
+            Boolean check = orderService.isRoomAvailable(roomId, start, end);
+            return ResponseEntity.ok(ApiResponse.success("查詢成功", check));
         } catch (Exception e) {
             e.printStackTrace();
         }
-        return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        return ResponseEntity.badRequest().body(ApiResponse.error(400, "查詢失敗"));
     }
 
     @GetMapping("/all")
-    public ResponseEntity<List<List<OrderDto>>> getUserAllOrders(@RequestHeader("Authorization") String jwt) {
+    public ResponseEntity<ApiResponse<List<List<OrderDto>>>> getUserAllOrders(@RequestHeader("Authorization") String jwt) {
 
-        Long userId = objectMapper.convertValue(authInterface.findUserIdByJwt(jwt).getBody().getData(), Long.class);
+        Long userId = authInterface.findUserIdByJwt(jwt).getBody().getData();
 
         List<List<OrderDto>> allOrders = new ArrayList<>();
         try {
@@ -98,91 +104,108 @@ public class OrderController {
             List<OrderDto> disannulOrders = orderService.getUserDisannulOrder(userId);
             List<OrderDto> finishedOrders = orderService.getUserFinishedOrder(userId);
             List<OrderDto> canceledOrders = orderService.getUserCanceledOrder(userId);
-            System.out.println("使用者" + userId + "有" + validOrders.size() + "筆有效訂單,有" + pendingOrders.size() + "筆等待中訂單,有" + disannulOrders.size() + "筆欲取消訂單,有" + canceledOrders.size() + "筆取消訂單,有" + finishedOrders.size() + "筆完成訂單");
             allOrders.add(validOrders);
             allOrders.add(pendingOrders);
             allOrders.add(disannulOrders);
             allOrders.add(finishedOrders);
             allOrders.add(canceledOrders);
-            return new ResponseEntity<>(allOrders, HttpStatus.OK);
+            return ResponseEntity.ok(ApiResponse.success("成功查詢使用者訂單", allOrders));
         } catch (Exception e) {
             e.printStackTrace();
         }
-        return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
+        return ResponseEntity.badRequest().body(ApiResponse.error(400, "查詢失敗，請重新嘗試"));
     }
 
     @PostMapping("/availability")
-    public ResponseEntity<List<Long>> checkHotelAvailableRooms(@RequestBody CheckRoomAvailableRequest request) {
+    public ResponseEntity<ApiResponse<List<Long>>> checkHotelAvailableRooms(@RequestBody CheckRoomAvailableRequest request) {
         try {
             List<Long> hotelRoomAvailable = orderService.filterHotelUnavailableRoom(request.getRoomIds(), request.getStart(), request.getEnd());
-            return new ResponseEntity<>(hotelRoomAvailable, HttpStatus.OK);
-
+            return ResponseEntity.ok(ApiResponse.success("查詢成功", hotelRoomAvailable));
         } catch (Exception e) {
             e.printStackTrace();
         }
-        return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        return ResponseEntity.badRequest().body(ApiResponse.error(400, "查詢失敗，請重新嘗試"));
     }
 
     @PostMapping("/hotel")
-    public ResponseEntity<Integer> getHotelOrderCount(@RequestBody List<Long> roomIds) {
+    public ResponseEntity<ApiResponse<Integer>> getHotelOrderCount(@RequestBody List<Long> roomIds) {
         try {
             System.out.println("我要搜尋訂單數量");
             Integer orders = orderService.getOrderCountByRoomList(roomIds);
             if (orders != null) {
-                return new ResponseEntity<>(orders, HttpStatus.OK);
+                return ResponseEntity.ok(ApiResponse.success("查詢成功", orders));
             }
-        } catch (RuntimeException e) {
-            throw new RuntimeException("搜尋訂單失敗" + roomIds);
+        } catch (Exception e) {
+            e.printStackTrace();
         }
-        return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
+        return ResponseEntity.badRequest().body(ApiResponse.error(400, "查詢失敗，請重新嘗試"));
     }
 
     @PutMapping("/cancel/{orderId}")
-    public ResponseEntity<Orders> cancelOrder(@RequestHeader("Authorization") String jwt, @PathVariable Long orderId) {
+    public ResponseEntity<ApiResponse<String>> cancelOrder(@RequestHeader("Authorization") String jwt, @PathVariable Long orderId) {
         try {
-            System.out.println("準備刪除訂單");
-            Orders order = orderService.cancelOrder(orderId);
-            return new ResponseEntity<>(order, HttpStatus.OK);
+            Long userId = authInterface.findUserIdByJwt(jwt).getBody().getData();
+            Orders order = orderService.findOrderByOrderId(orderId);
+            if (order.getUserId() != userId) {
+                log.error("(cancelOrder)使用者"+userId+"嘗試執行非法操作");
+                return ResponseEntity.badRequest().body(ApiResponse.error(400, "無權限執行此操作"));
+            }
+
+            orderService.cancelOrder(orderId);
+            return ResponseEntity.ok(ApiResponse.success("訂單取消成功",null));
         } catch (RuntimeException e) {
             e.printStackTrace();
         }
-        return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
+        return ResponseEntity.badRequest().body(ApiResponse.error(400,"操作失敗，請重新嘗試"));
     }
 
     @PutMapping("/wantCancel/{orderId}")
-    public ResponseEntity<Orders> wantCancelOrder(@RequestHeader("Authorization") String jwt, @PathVariable Long orderId) {
+    public ResponseEntity<ApiResponse<Orders>> wantCancelOrder(@RequestHeader("Authorization") String jwt, @PathVariable Long orderId) {
         try {
-            System.out.println("準備刪除訂單");
-            Orders order = orderService.wantCancelOrder(orderId);
-            return new ResponseEntity<>(order, HttpStatus.OK);
+            Long userId = authInterface.findUserIdByJwt(jwt).getBody().getData();
+            Orders order = orderService.findOrderByOrderId(orderId);
+            if (order.getUserId() != userId) {
+                log.error("(wantCancelOrder)使用者"+userId+"嘗試執行非法操作");
+                return ResponseEntity.badRequest().body(ApiResponse.error(400, "無權限執行此操作"));
+            }
+            orderService.wantCancelOrder(orderId);
+            return ResponseEntity.ok(ApiResponse.success("取消訂單需求成功提出",null));
         } catch (RuntimeException e) {
             e.printStackTrace();
         }
-        return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
+        return ResponseEntity.badRequest().body(ApiResponse.error(400,"操作失敗，請重新嘗試"));
     }
 
     @GetMapping("/{orderId}")
-    public ResponseEntity<Orders> getOrderData(@PathVariable Long orderId) {
+    public ResponseEntity<ApiResponse<Orders>> getOrderData(@PathVariable Long orderId) {
         try {
-            System.out.println("我要找這筆訂單" + orderId);
             Orders order = orderService.findOrderByOrderId(orderId);
-            return new ResponseEntity<>(order, HttpStatus.OK);
+            System.out.println("我要找這筆訂單" + orderId);
+            return ResponseEntity.ok(ApiResponse.success("查詢成功",order));
         } catch (RuntimeException e) {
             e.printStackTrace();
         }
-        return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
+        return ResponseEntity.badRequest().body(ApiResponse.error(400,"操作失敗，請重新嘗試"));
     }
 
     @PutMapping("/{orderId}/commented/{status}")
-    public ResponseEntity<String> updateOrderCommentStatus(@RequestHeader("Authorization") String jwt,@PathVariable Long orderId,@PathVariable Boolean status) {
-        try {
-            Long userId = objectMapper.convertValue(authInterface.findUserIdByJwt(jwt).getBody().getData(), Long.class);
-            System.out.println("我要改這筆訂單" + orderId);
-            if (orderService.updateOrderCommentStatus(userId, orderId, status))
-                return new ResponseEntity<>("修改成功", HttpStatus.OK);
-        } catch (RuntimeException e) {
+    public ResponseEntity<ApiResponse<String>> updateOrderCommentStatus(@RequestHeader("Authorization") String jwt,
+                                                                        @PathVariable Long orderId,
+                                                                        @PathVariable Boolean status){
+        try{
+            Long userId = authInterface.findUserIdByJwt(jwt).getBody().getData();
+            Orders order = orderService.findOrderByOrderId(orderId);
+            if(userId!=order.getUserId()) {
+                log.error("使用者{}嘗試修改訂單{}",userId,order.getId());
+                return ResponseEntity.badRequest().body(ApiResponse.error(400, "無權限執行此操作"));
+            }
+            orderService.updateOrderCommentStatus(order,status);
+            return ResponseEntity.ok(ApiResponse.success("評論狀態已修改為"+status,null));
+        } catch (Exception e) {
             e.printStackTrace();
+            return ResponseEntity.badRequest().body(ApiResponse.error(400,e.getMessage()));
         }
-        return new ResponseEntity<>("無法修改，非該評論使用者", HttpStatus.BAD_REQUEST);
     }
+
+
 }
